+++ test/marskal/marskal_core_ext_my_string_test.rb
@@ -5,7 +5,7 @@ require 'marskal/core_ext/string'
 # This Section Tests all the Marskal::CoreExt +String+ extensions
 #
 # ==== History
-# * <tt>Created: 2016-10-2016</tt> <b>Mike Urban</b> <mike@marskalgroup.com>
+# * <tt>Created: 2016-12-10</tt> <b>Mike Urban</b> <mike@marskalgroup.com>
 #
 # ---
 describe 'Marskal::CoreExt::MyString' do